Rename loader modules
